agrege productos, modifique el sistema de catalogo anterior y quedo funcional el sistema de filament

+++ database/migrations/2025_05_26_182129_add_inventory_count_to_products_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-        $table->integer('inventory_count')->default(0)->after('price');
-    });
+        if (!Schema::hasColumn('products', 'inventory_count')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->integer('inventory_count')->default(0)->after('price');
+            });
+        }
     }
 
     /**
@@ -21,9 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-        $table->dropColumn('inventory_count');
-    });
-
+        if (Schema::hasColumn('products', 'inventory_count')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('inventory_count');
+            });
+        }
     }
 };
